Added dialogs subfolder
